+++ demoApp/iosApp/iosApp/ComposeRootView.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+import sharedUI
+
+struct ComposeRootView: UIViewControllerRepresentable {
+    func makeUIViewController(context: Context) -> UIViewController {
+        MainViewControllerKt.MainViewController()
+    }
+
+    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {}
+}