+++ library/core/src/iosMain/kotlin/com/lib/lokdroid/data/default_implementation/TagProvider.kt
@@ -1,5 +1,0 @@
-package com.lib.lokdroid.data.default_implementation
-
-actual object TagProvider {
-    actual fun getTag(): String = "LoKdroidIos"
-}